app/requirements.txt
importしたものは
random
です
flet-pyodide
